Fix: fix timeout tipe as int
Documentation and Annotations require timeout to be a float, but the API
requires an integer.

Fixes: https://github.com/containers/podman-py/issues/534

+++ podman/domain/pods.py
@@ -5,7 +5,7 @@ from typing import Any, Optional, Union
 
 from podman.domain.manager import PodmanResource
 
-_Timeout = Union[None, float, tuple[float, float], tuple[float, None]]
+_Timeout = Union[None, int, tuple[int, int], tuple[int, None]]
 
 logger = logging.getLogger("podman.pods")
